resell

+++ app/Http/Controllers/ApiFront/ApiProposalController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\ApiFront;
 
+use App\Events\NewProposal;
 use App\Models\TypesOfJobs;
 use http\Env\Response;
 use Illuminate\Http\Request;
@@ -18,6 +19,23 @@ use PDF;
 class ApiProposalController extends Controller
 {
 
+    public function store(Request $request)
+    {
+
+        $proposal = $request->only('proposal')['proposal'];
+
+        if(auth()->user()->isPartner()){
+            $proposal['resell'] = 1;
+        }
+
+        $proposal['user_id'] = auth()->user()->id;
+        $proposal['additional_info'] = $request->only('additional_info')['additional_info'];
+
+        event(new NewProposal(Proposal::create($proposal)));
+
+        return response()->json(['redirect_url'=>route('partner.cabinet','antrag-von-firma/verkaufe')]);
+    }
+
     /**
      * Show proposals client.
      *
@@ -31,6 +49,35 @@ class ApiProposalController extends Controller
         $proposals = auth()->user()->getProposalsByTypeJob($typeJobsId)->get();
 
         return new ProposalCollection($proposals);
+    }
+
+
+    /**
+     * Show proposals resell.
+     *
+     */
+    public function resellList()
+    {
+        $typeJobId = request()->get('type_job_id', null);
+        $regionId = request()->get('region_id', null);
+        $date = request()->get('date', null);
+
+        $proposals = new Proposal();
+        $proposals = $proposals->resellNoPay();
+
+        if($date){
+            $proposals = $proposals->whereDateStart($date);
+        }
+
+        if($regionId){
+            $proposals = $proposals->whereIn('region_id',$regionId);
+        }
+
+        if($typeJobId){
+            $proposals = $proposals->whereIn('type_job_id',$typeJobId);
+        }
+
+        return new ProposalCollection($proposals->paginate(10));
     }
 
     /**
@@ -47,6 +94,12 @@ class ApiProposalController extends Controller
             case 'rejected':
                 $proposals = auth()->user()->getProposalsByStatusCabinet(2)->get();
                 break;
+            case 'resell':
+                $proposals = auth()->user()->getResellProposals()->get();
+                break;
+            case 'resell-accepted':
+                $proposals = auth()->user()->getResellProposalsAccept(1)->get();
+                break;
             default:
                 $proposals = auth()->user()->getNewProposals()->get();
         }
@@ -62,7 +115,7 @@ class ApiProposalController extends Controller
     {
         return response()->json([
             'data'=>[
-                'all_proposals' => auth()->user()->getCountProposalsCabinet(),
+                'all_proposals' => auth()->user()->getProposalsByStatusCabinet(0)->count() + auth()->user()->getProposalsByStatusCabinet(1)->count() ,
                 'new_proposals' => auth()->user()->getProposalsByStatusCabinet(0)->count()
             ]
         ], 200);
@@ -111,8 +164,19 @@ class ApiProposalController extends Controller
      */
     public function processProposals(Request $request, Proposal $proposal)
     {
+
+        if($proposal->resell == 1){
+            return $this->resellProcess($proposal);
+        }else{
+            return $this->process($proposal);
+        }
+
+    }
+
+    private function process(Proposal $proposal){
         $authUser = auth()->user();
-        $status = $request->action == 'accepted' ? 1 : 2;
+
+        $status = request()->action == 'accepted' ? 1 : 2;
 
         $proposalToPartner = ProposalToPartner::whereUserId($authUser->id)
             ->where('proposal_id',$proposal->id)
@@ -120,47 +184,14 @@ class ApiProposalController extends Controller
             ->first();
 
         if (is_null($proposalToPartner)){
-            return response()->json(['data'=>['result'=>false,'message'=>'Not your request']]);
+            return response()->json(['data'=>['result'=>false,'message'=>'Your Not owner request']]);
         }
 
-        switch ($proposal['type_job_id']) {
-            case 1:
-                $price = Setting::getByKey('system.setting.cost_transfer');
-                break;
-            case 2:
-                $price = Setting::getByKey('system.setting.cost_cleaning');
-                break;
-            case 3:
-                $price = Setting::getByKey('system.setting.cost_transfer_cleaning');
-                break;
-            case 4:
-                $price = Setting::getByKey('system.setting.cost_malar');
-                break;
-            case 5:
-                $price = Setting::getByKey('system.setting.bodenleger');
-                break;
-            case 6:
-                $price = Setting::getByKey('system.setting.cost_malar');
-                break;
-            case 7:
-                $price = Setting::getByKey('system.setting.cost_malar');
-                break;
-            case 8:
-                $price = Setting::getByKey('system.setting.cost_malar');
-                break;
-            case 9:
-                $price = Setting::getByKey('system.setting.cost_malar');
-                break;
-            default:
-                $price = 0;
-                Log::info('Wrong Job Type on pay: '.$proposal['type_job_id']);
-        }
+        $price = Setting::getByKey('system.setting.cost_'.config('services.types_jobs.'.$proposal->type_job_id));
 
 
         if($status == 1) {
             if ($authUser->status != 1 && $authUser->status != 2) {
-
-
                 if($authUser->status_pay == 1){
 
                     if($authUser->coins - $price >= 0) {
@@ -186,13 +217,16 @@ class ApiProposalController extends Controller
                         ->where('status', 0)
                         ->forceDelete();
                 }
+
                 return response()->json(['data'=>['result'=>true]]);
-              //  return redirect()->route('partner.getAcceptedRequests');
+
 
             }else{
                 return response()->json(['data'=>['result'=>true,'message'=>__('front.no_paid_invoice')]]);
             }
+            return response()->json(['data'=>['result'=>true,'redirect_url'=>route('partner.cabinet','anfragen/angenommene')]]);
         }
+
         $proposalToPartner->status = $status;
         $proposalToPartner->save();
 
@@ -201,17 +235,58 @@ class ApiProposalController extends Controller
     /**
      * download proposal.
      * @param Proposal $proposal
+     */
+    private function resellProcess(Proposal $proposal){
+        $authUser = auth()->user();
+
+        ProposalToPartner::create([
+            'user_id'=>auth()->ID(),
+            'proposal_id'=>$proposal->id,
+            'status'=> 1,
+        ]);
+
+        $price = (int)$proposal->price;
+
+        if($authUser->status_pay == 1){
+
+            if($authUser->coins - $price >= 0) {
+                $authUser->coins = $authUser->coins - $price;
+                $authUser->save();
+            }else{
+                return response()->json(['data'=>['result'=>false ,'no_coin'=>__('front.no_coin')]]);
+            }
+        }
+
+        Log::info('Partner ID: ' . $authUser->id . ', Name: ' . $authUser->name . ' PAY Proposal Resell ID: ' . $proposal->id.' Price: '.$price);
+        event(new ProposalAccepted($proposal));
+        Log::info('----DONE----');
+
+        $proposal->payed = 1;
+        $proposal->save();
+
+        return response()->json(['data'=>['result'=>true]]);
+    }
+
+    /**
+     * download proposal.
+     * @param Proposal $proposal
      * @return PDF
      */
     public function download(Proposal $proposal){
 
-        $cost = Setting::getByKey('system.setting.cost_'.config('services.types_jobs.'.$proposal->type_job_id));
-
+        if($proposal->resell == 1){
+            $cost = (int)$proposal->price;
+        }else{
+            $cost = Setting::getByKey('system.setting.cost_'.config('services.types_jobs.'.$proposal->type_job_id));
+        }
         $pdf = PDF::loadView('front.partner.proposalPDF',compact(['proposal','cost']));
 
         return $pdf->download('anfragen_'. $proposal->date_start->format('Y-m-d') .'.pdf');
     }
-
+    /**
+     * List prices request.
+     * @return Response
+     */
     public function prices(){
         $typesJobs = TypesOfJobs::all();
 
@@ -229,11 +304,18 @@ class ApiProposalController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy(Proposal $proposal){
-        event(new ProposalDelete($proposal));
 
+        $proposals = auth()->user()->getProposalsByClient()->whereId($proposal->id)->get();
+        if($proposals->isEmpty()){
+            return response()->json(['data'=>['result'=>false,'message'=> 'your not owner']]);
+        }
+
+        event(new ProposalDelete($proposal));
         $proposal->getReviews()->delete();
         $proposal->getReceivedInvitation()->forceDelete();
         $proposal->forceDelete();
+
+
         return response()->json(['data'=>['result'=>true]]);
     }
